move mapping protocol to rustpython_vm::protocol
