+++ vm/src/builtins/dict.rs
@@ -4,7 +4,7 @@ use crate::{
     common::ascii,
     dictdatatype::{self, DictKey},
     function::{ArgIterable, FuncArgs, KwArgs, OptionalArg},
-    protocol::PyIterReturn,
+    protocol::{PyIterReturn, PyMapping},
     slots::{
         AsMapping, Comparable, Hashable, Iterable, IteratorIterable, PyComparisonOp, SlotIterator,
         Unhashable,
@@ -13,7 +13,7 @@ use crate::{
     IdProtocol, IntoPyObject, ItemProtocol,
     PyArithmeticValue::*,
     PyAttributes, PyClassDef, PyClassImpl, PyComparisonValue, PyContext, PyObjectRef, PyRef,
-    PyResult, PyValue, TryFromBorrowedObject, TypeProtocol,
+    PyResult, PyValue, TypeProtocol,
 };
 use crossbeam_utils::atomic::AtomicCell;
 use std::fmt;
@@ -934,37 +934,4 @@ pub(crate) fn init(context: &PyContext) {
     PyDictItems::extend_class(context, &context.types.dict_items_type);
     PyDictItemIterator::extend_class(context, &context.types.dict_itemiterator_type);
     PyDictReverseItemIterator::extend_class(context, &context.types.dict_reverseitemiterator_type);
-}
-
-#[allow(clippy::type_complexity)]
-pub struct PyMapping {
-    pub length: Option<fn(PyObjectRef, &VirtualMachine) -> PyResult<usize>>,
-    pub subscript: Option<fn(PyObjectRef, PyObjectRef, &VirtualMachine) -> PyResult>,
-    pub ass_subscript:
-        Option<fn(PyObjectRef, PyObjectRef, Option<PyObjectRef>, &VirtualMachine) -> PyResult<()>>,
-}
-
-impl PyMapping {
-    pub fn check(cls: &PyObjectRef, vm: &VirtualMachine) -> bool {
-        if let Ok(mapping) = PyMapping::try_from_borrowed_object(vm, cls) {
-            mapping.subscript.is_some()
-        } else {
-            false
-        }
-    }
-}
-
-impl TryFromBorrowedObject for PyMapping {
-    fn try_from_borrowed_object(vm: &VirtualMachine, obj: &PyObjectRef) -> PyResult<Self> {
-        let obj_cls = obj.class();
-        for cls in obj_cls.iter_mro() {
-            if let Some(f) = cls.slots.as_mapping.load() {
-                return f(obj, vm);
-            }
-        }
-        Err(vm.new_type_error(format!(
-            "a dict-like object is required, not '{}'",
-            obj_cls.name()
-        )))
-    }
 }
