+++ vm/src/builtins/pytype.rs
@@ -4,9 +4,8 @@ use super::{
 };
 use crate::common::{ascii, lock::PyRwLock};
 use crate::{
-    builtins::dict::PyMapping,
     function::{FuncArgs, KwArgs, OptionalArg},
-    protocol::PyIterReturn,
+    protocol::{PyIterReturn, PyMapping},
     slots::{self, Callable, PyTypeFlags, PyTypeSlots, SlotGetattro, SlotSetattro},
     utils::Either,
     IdProtocol, PyAttributes, PyClassImpl, PyContext, PyLease, PyObjectRef, PyRef, PyResult,
