+++ vm/src/builtins/tuple.rs
@@ -3,11 +3,10 @@ use super::{
     iter::IterStatus::{self, Active, Exhausted},
     PyInt, PyTypeRef,
 };
-use crate::builtins::dict::PyMapping;
 use crate::common::hash::PyHash;
 use crate::{
     function::OptionalArg,
-    protocol::PyIterReturn,
+    protocol::{PyIterReturn, PyMapping},
     sequence::{self, SimpleSeq},
     sliceable::PySliceableSequence,
     slots::{
