+++ vm/src/protocol/mod.rs
@@ -1,5 +1,7 @@
 mod buffer;
 mod iter;
+mod mapping;
 
 pub use buffer::{BufferInternal, BufferOptions, PyBuffer, ResizeGuard};
 pub use iter::{PyIter, PyIterReturn};
+pub(crate) use mapping::PyMapping;
