+++ vm/src/pyobject.rs
@@ -10,7 +10,6 @@ use crate::{
         builtinfunc::PyNativeFuncDef,
         bytearray, bytes,
         code::{self, PyCode},
-        dict::PyMapping,
         getset::{IntoPyGetterFunc, IntoPySetterFunc, PyGetSet},
         namespace::PyNamespace,
         object, pystr,
@@ -21,6 +20,7 @@ use crate::{
     dictdatatype::Dict,
     exceptions,
     function::{IntoFuncArgs, IntoPyNativeFunc},
+    protocol::PyMapping,
     slots::{PyTypeFlags, PyTypeSlots},
     types::{create_type_with_slots, TypeZoo},
     VirtualMachine,
